requirements.txt
aiofiles
aiohttp
aiologger
argparse
pillow
requests
tqdm
uvloop; sys_platform != 'win32'
